// frontend/src/app/api/models/admin-match.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MatchStatus } from '../models/match-status';
import { Score } from '../models/score';
import { Team } from '../models/team';
export interface AdminMatch {
  awayTeam: Team;
  group: string;
  homeTeam: Team;
  id: number;
  kickoffAt: string;
  matchday: number;
  predictionsCount: number;
  result?: Score | null;
  resultRegisteredAt?: string | null;
  status: MatchStatus;
  version: number;
}